requirements.txt
Django==3.1.1
django-extended-choices==1.3.3


# api
djangorestframework==3.12.1
markdown==3.3.3
django-filter==2.4.0
django-treebeard
psycopg2-binary
django-allauth
django-cors-headers

# frontend
django-webpack-loader==0.7.0

# data
requests==2.24.0
pandas
numpy

# for prod
gunicorn==20.0.4

# dev
pytest-django
django_debug_toolbar
django-extensions
ipython
pyflakes

# devops
python-dotenv==0.14.0
# fab-classic is a Fabric fork with py3 support, but faithful to the old API
fab-classic==1.18.1

